Core is Done - Missing features: Spotify and OpenWeather Api integration - Interesting Features: HATOAS and SWAGGER

+++ src/main/java/com/ifood/challenge/core/playlistsuggester/business/WeatherMusicGenreExpression.java
@@ -3,13 +3,13 @@ package com.ifood.challenge.core.playlistsuggester.business;
 import java.io.Serializable;
 
 public class WeatherMusicGenreExpression implements Serializable {
-  private Long temperature;
+  private Integer temperature;
 
-  public WeatherMusicGenreExpression(Long temperature) {
+  public WeatherMusicGenreExpression(Integer temperature) {
     this.temperature = temperature;
   }
 
-  public Long getTemperature() {
+  public Integer getTemperature() {
     return temperature;
   }
 }
